feat(payment): 打赏与课程购买支付成功联动等级经验及系统通知
- 支付宝回调确认支付成功后，按实际支付金额给付款用户加逐日等级经验（金额即经验值，支持小数），并发送系统通知站内信
- 新增 PaymentRewardService 聚合联动，任一联动失败仅记日志不影响支付主流程
- 逐日经验字段 int 升级 decimal(10,2)，实体改 BigDecimal，新增迁移脚本 alter_score_decimal.sql
- 新增文章打赏功能（打赏卡片/弹窗/流水表 ap_article_tip_order、ap_article_tip_record）
- 完成课程订单详情页、支付回跳地址修复、课程列表新开标签等前端

+++ heima-leadnews/heima-leadnews-model/src/main/java/com/heima/model/level/pojos/ApUserLevel.java
@@ -7,6 +7,7 @@ import com.baomidou.mybatisplus.annotation.TableName;
 import lombok.Data;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.util.Date;
 
 @Data
@@ -20,7 +21,7 @@ public class ApUserLevel implements Serializable {
     private Long userId;
 
     @TableField("daily_score")
-    private Integer dailyScore;
+    private BigDecimal dailyScore;
 
     @TableField("daily_level")
     private Integer dailyLevel;
@@ -32,7 +33,7 @@ public class ApUserLevel implements Serializable {
     private Integer powerLevel;
 
     @TableField("daily_score_today")
-    private Integer dailyScoreToday;
+    private BigDecimal dailyScoreToday;
 
     @TableField("power_value_today")
     private Integer powerValueToday;
